+++ crates/database/src/vm_database.rs
@@ -1,24 +1,9 @@
 use crate::Error as DatabaseError;
 use anyhow::anyhow;
-use fuel_core_storage::{
-    iter::IterDirection,
-    not_found,
-    tables::{
-        ContractsAssets,
-        ContractsState,
-    },
-    ContractsAssetsStorage,
-    ContractsStateKey,
-    Error as StorageError,
-    Mappable,
-    MerkleRoot,
-    MerkleRootStorage,
-    StorageAsMut,
-    StorageInspect,
-    StorageMutate,
-    StorageRead,
-    StorageSize,
-};
+use fuel_core_storage::{iter::IterDirection, not_found, tables::{
+    ContractsAssets,
+    ContractsState,
+}, ContractsAssetsStorage, ContractsStateKey, Error as StorageError, Mappable, MerkleRoot, MerkleRootStorage, StorageAsMut, StorageInspect, StorageMutate, StorageRead, StorageSize};
 use fuel_core_types::{
     blockchain::header::ConsensusHeader,
     fuel_tx::{
@@ -37,18 +22,15 @@ use fuel_core_types::{
 };
 use primitive_types::U256;
 use std::borrow::Cow;
+use fuel_core_types::blockchain::primitives::BlockId;
 
 use serde::de::DeserializeOwned;
 
-use fuel_core_executor::refs::{
-    ExecutorVmDatabase,
-    FuelBlockTrait,
-    FuelStateTrait,
-};
 use fuel_core_storage::tables::{
     ContractsInfo,
     ContractsRawCode,
 };
+use fuel_core_types::fuel_types::Nonce;
 
 /// Used to store metadata relevant during the execution of a transaction
 #[derive(Clone, Debug)]
@@ -182,6 +164,36 @@ impl<D> ContractsAssetsStorage for VmDatabase<D> where
 {
 }
 
+
+pub trait FuelBlockTrait {
+    type Error;
+
+    fn latest_height(&self) -> Result<BlockHeight, Self::Error>;
+    fn block_time(&self, height: &BlockHeight) -> Result<Tai64, Self::Error>;
+    fn get_block_id(&self, height: &BlockHeight) -> Result<Option<BlockId>, Self::Error>;
+}
+
+pub trait FuelStateTrait {
+    type Error;
+
+    fn init_contract_state<S: Iterator<Item = (Bytes32, Bytes32)>>(
+        &mut self,
+        contract_id: &ContractId,
+        slots: S,
+    ) -> Result<(), Self::Error>;
+}
+
+use fuel_core_storage::tables::SpentMessages;
+use fuel_core_storage::tables::Messages;
+
+pub trait MessageIsSpent:
+    StorageInspect<SpentMessages, Error = StorageError>
++   StorageInspect<Messages, Error = StorageError> {
+    type Error;
+
+    fn message_is_spent(&self, nonce: &Nonce) -> Result<bool, StorageError>;
+}
+
 pub trait DatabaseIteratorsTrait {
     fn iter_all_filtered_column<K, V, P, S>(
         &self,
@@ -196,16 +208,6 @@ pub trait DatabaseIteratorsTrait {
         S: AsRef<[u8]>;
 }
 
-impl<D> From<ExecutorVmDatabase<D>> for VmDatabase<D> {
-    fn from(muda: ExecutorVmDatabase<D>) -> Self {
-        VmDatabase {
-            current_block_height: muda.current_block_height,
-            current_timestamp: muda.current_timestamp,
-            coinbase: muda.coinbase,
-            database: muda.database,
-        }
-    }
-}
 
 impl<D> InterpreterStorage for VmDatabase<D>
 where
@@ -389,6 +391,7 @@ where
             Ok(Some(()))
         }
     }
+
 }
 
 #[cfg(test)]
