+++ crates/fuel-core/src/lib.rs
@@ -25,7 +25,6 @@ pub use fuel_core_types as types;
 
 pub mod coins_query;
 pub mod database;
-pub mod executor;
 pub mod model;
 #[cfg(all(feature = "p2p", feature = "test-helpers"))]
 pub mod p2p_test_helpers;
